JCRVLT-677 store attributes within "entries.xml"
This fixes the NPE being issued when trying to resolve conflicts
Convert tests to JUnit5
