+++ parent/pom.xml
@@ -529,6 +529,13 @@ Bundle-Category: jackrabbit
                 <artifactId>junit</artifactId>
                 <version>4.13.2</version>
             </dependency>
+            <dependency>
+                <groupId>org.junit</groupId>
+                <artifactId>junit-bom</artifactId>
+                <version>5.9.2</version>
+                <type>pom</type>
+                <scope>import</scope>
+            </dependency>
         </dependencies>
     </dependencyManagement>
